requirements.txt
certifi==2019.9.11
chardet==3.0.4
Click==7.0
Flask==1.1.1
Flask-SQLAlchemy==2.4.1
Flask-SSLify==0.1.5
idna==2.8
itsdangerous==1.1.0
Jinja2==2.10.3
lml==0.0.9
MarkupSafe==1.1.1
pyexcel-io==0.5.20
pyexcel-xls==0.5.8
python-Levenshtein==0.12.0
pytz==2019.3
requests==2.22.0
SQLAlchemy==1.3.11
urllib3==1.25.7
Werkzeug==0.16.0
xlrd==1.2.0
xlwt==1.3.0
